Make source_plugin a separate engine
