+++ userspace/libscap/scap-int.h
@@ -197,28 +197,6 @@ struct scap
 
 	bool syscalls_of_interest[SYSCALL_TABLE_SIZE];
 
-	//
-	// Plugin-related state
-	//
-	scap_source_plugin* m_input_plugin;
-	uint8_t* m_input_plugin_evt_storage;
-	uint32_t m_input_plugin_evt_storage_len;
-
-	// The number of items held in batch_evts
-	uint32_t m_input_plugin_batch_nevts;
-
-	// A set of events returned from next_batch. The array is
-	// allocated and must be free()d when done.
-	ss_plugin_event* m_input_plugin_batch_evts;
-
-	// The current position into the above arrays (0-indexed),
-	// reflecting how many of the above items have been returned
-	// via a call to next().
-	uint32_t m_input_plugin_batch_idx;
-
-	// The return value from the last call to next_batch().
-	ss_plugin_rc m_input_plugin_last_batch_res;
-
 	// API version supported by the driver
 	// If the API version is unavailable for whatever reason,
 	// it's equivalent to version 0.0.0
